Restricting subsets due to locked projects

+++ geokey/subsets/views.py
@@ -8,90 +8,109 @@ from django.contrib import messages
 from braces.views import LoginRequiredMixin
 
 from geokey.core.decorators import handle_exceptions_for_admin
-from geokey.projects.views import ProjectContext
 from geokey.projects.models import Project
+from geokey.projects.views import ProjectContext
 
 from .models import Subset
 
 
 class SubsetOverview(LoginRequiredMixin, ProjectContext, TemplateView):
+
     """
-    Displays the list of subsets in a project.
+    Displays the list of subsets in the project.
     """
-    template_name = 'subsets/subsets_overview.html'
+    template_name = 'subsets/subset_overview.html'
 
 
 class SubsetCreate(LoginRequiredMixin, ProjectContext, TemplateView):
+
     """
     Provides the form to create a new subset.
     """
-    template_name = 'subsets/subsets_create.html'
+    template_name = 'subsets/subset_create.html'
 
     def post(self, request, project_id):
         """
-        Creates the subset based on the data entered by the user
+        Creates the subset based on the data entered by the user.
 
-        Parameter
-        ---------
+        Parameters
+        ----------
         request : django.http.HttpRequest
-            Object representing the request.
+            Object representing the request
         project_id : int
-            identifies the project in the data base
+            Identifies the project in the database
 
         Returns
         -------
         django.http.HttpResponseRedirect
-            to SubsetData
+            Redirects to subset data if subset is created, project overview if
+            project is locked
         django.http.HttpResponse
-            Rendered template, if an error occured
+            Rendered template, if project does not exist
         """
+
         data = request.POST
         context = self.get_context_data(project_id)
-
         project = context.get('project')
 
         if project:
-            subset = Subset.objects.create(
-                name=strip_tags(data.get('name')),
-                description=strip_tags(data.get('description')),
-                creator=request.user,
-                project=project
-            )
-            messages.success(self.request, "The subset has been created.")
-            return redirect(
-                'admin:subset_data',
-                project_id=project.id,
-                subset_id=subset.id
-            )
+            if project.islocked:
+                messages.error(
+                    self.request,
+                    'The project is locked. New subsets cannot be created.'
+                )
+                return redirect(
+                    'admin:project_overview',
+                    project_id=project_id
+                )
+            else:
+                subset = Subset.objects.create(
+                    name=strip_tags(data.get('name')),
+                    description=strip_tags(data.get('description')),
+                    creator=request.user,
+                    project=project
+                )
+
+                messages.success(self.request, 'The subset has been created.')
+                return redirect(
+                    'admin:subset_data',
+                    project_id=project_id,
+                    subset_id=subset.id
+                )
         else:
             return self.render_to_response(context)
 
 
 class SubsetContext(object):
+
     """
     Mixin that provides the context to render templates. The context contains
-    a Subset instance based on project_id and subset_id.
+    a subset instance based on project_id and subset_id
     """
+
     @handle_exceptions_for_admin
     def get_context_data(self, project_id, subset_id, *args, **kwargs):
         """
-        Returns the context containing the project instance.
+        Returns the context containing the project and subset instances.
 
         Parameters
         ----------
         project_id : int
-            identifies the project in the data base
+            Identifies the project in the database
         subset_id : int
-            identifies the subset in the data base
+            Identifies the subset in the database
 
         Returns
         -------
         dict
+            Context
         """
+
         project = Project.objects.as_admin(self.request.user, project_id)
         subset = project.subsets.get(pk=subset_id)
 
         return super(SubsetContext, self).get_context_data(
+            project=project,
             subset=subset,
             *args,
             **kwargs
@@ -99,48 +118,53 @@ class SubsetContext(object):
 
 
 class SubsetSettings(LoginRequiredMixin, SubsetContext, TemplateView):
+
     """
     Provides the form to update the subset settings.
     """
-    template_name = 'subsets/subsets_settings.html'
+    template_name = 'subsets/subset_settings.html'
 
     def post(self, request, project_id, subset_id):
         """
-        Updates the subset based on the data entered by the user
+        Updates the subset based on the data entered by the user.
 
         Parameter
         ---------
         request : django.http.HttpRequest
-            Object representing the request.
+            Object representing the request
         project_id : int
-            identifies the project in the data base
+            Identifies the project in the database
         subset_id : int
-            identifies the subset in the data base
+            Identifies the subset in the database
 
         Returns
         -------
         django.http.HttpResponse
-            Rendered template
+            Rendered template when subset updated
+        django.http.HttpResponse
+            Rendered template, if project or subset does not exist
         """
+
+        data = request.POST
         context = self.get_context_data(project_id, subset_id)
         subset = context.get('subset')
 
         if subset:
-            data = request.POST
             subset.name = strip_tags(data.get('name'))
             subset.description = strip_tags(data.get('description'))
             subset.save()
 
-            messages.success(self.request, "The subset has been updated.")
+            messages.success(self.request, 'The subset has been updated.')
 
         return self.render_to_response(context)
 
 
 class SubsetData(LoginRequiredMixin, SubsetContext, TemplateView):
+
     """
-    Provides the form to change the filter settings for the subset
+    Provides the form to change the filter settings for the subset.
     """
-    template_name = 'subsets/subsets_data.html'
+    template_name = 'subsets/subset_data.html'
 
     def post(self, request, project_id, subset_id):
         """
@@ -149,63 +173,90 @@ class SubsetData(LoginRequiredMixin, SubsetContext, TemplateView):
         Parameter
         ---------
         request : django.http.HttpRequest
-            Object representing the request.
+            Object representing the request
         project_id : int
-            identifies the project in the data base
+            Identifies the project in the database
         subset_id : int
-            identifies the subset in the data base
+            Identifies the subset in the database
 
         Returns
         -------
         django.http.HttpResponse
-            Rendered template
+            Rendered template when subset updated
+        django.http.HttpResponse
+            Rendered template, if project or subset does not exist
         """
+
+        data = request.POST
         context = self.get_context_data(project_id, subset_id)
         subset = context.get('subset')
 
         if subset:
-            data = request.POST
-            if data['filters'] != '-1':
-                subset.filters = json_loads(data['filters'])
-                subset.save()
+            if subset.project.islocked:
+                messages.error(
+                    self.request,
+                    'The project is locked. Subsets cannot be edited.'
+                )
+            else:
+                if data['filters'] != '-1':
+                    subset.filters = json_loads(data['filters'])
+                    subset.save()
 
-            messages.success(self.request, "The subset has been updated.")
+                    messages.success(
+                        self.request,
+                        'The subset has been updated.'
+                    )
 
         return self.render_to_response(context)
 
 
 class SubsetDelete(LoginRequiredMixin, SubsetContext, TemplateView):
+
     """
-    Deletes a subset
+    Deletes the subset.
     """
     template_name = 'base.html'
 
     def get(self, request, project_id, subset_id):
         """
-        Deletes the subset
+        Deletes the subset.
 
         Parameter
         ---------
         request : django.http.HttpRequest
-            Object representing the request.
+            Object representing the request
         project_id : int
-            identifies the project in the data base
+            Identifies the project in the database
         subset_id : int
-            identifies the subset in the data base
+            Identifies the subset in the database
 
         Returns
         -------
         django.http.HttpResponseRedirect
-            to SubsetOverview
+            Redirects to subset list if subset is deleted, subset settings if
+            project is locked
         django.http.HttpResponse
-            Rendered template, if an error has occured
+            Rendered template, if project or subset does not exist
         """
+
         context = self.get_context_data(project_id, subset_id)
         subset = context.get('subset')
 
         if subset:
-            subset.delete()
-            messages.success(self.request, "The subset has been deleted.")
-            return redirect('admin:subset_list', project_id=project_id)
+            if subset.project.islocked:
+                messages.error(
+                    self.request,
+                    'The project is locked. Subsets cannot be deleted.'
+                )
+                return redirect(
+                    'admin:subset_settings',
+                    project_id=project_id,
+                    subset_id=subset_id
+                )
+            else:
+                subset.delete()
+
+                messages.success(self.request, 'The subset has been deleted.')
+                return redirect('admin:subset_list', project_id=project_id)
 
         return self.render_to_response(context)
